power_service/requirements.txt
flask==0.12.2
gunicorn
python-dateutil